fix: fmt

+++ test/invariant/fuzz/Setup.t.sol
@@ -14,12 +14,7 @@ import {Utils} from "./helpers/Utils.t.sol";
 import {FuzzERC20, ERC20} from "./helpers/FuzzERC20.sol";
 
 contract Setup is Actors {
-    address[] DEFAULT_MEDUSA_SENDER = [
-        address(0x10000),
-        address(0x20000),
-        address(0x30000),
-        address(0x40000)
-    ];
+    address[] DEFAULT_MEDUSA_SENDER = [address(0x10000), address(0x20000), address(0x30000), address(0x40000)];
 
     uint256 percentFee;
     uint256 baseFee;
@@ -46,24 +41,13 @@ contract Setup is Actors {
         registry = new Registry();
 
         // Deploy the proxy, pointing to the implementation
-        TransparentUpgradeableProxy proxy = new TransparentUpgradeableProxy(
-            implementation,
-            proxyOwner,
-            ""
-        );
+        TransparentUpgradeableProxy proxy = new TransparentUpgradeableProxy(implementation, proxyOwner, "");
 
         allo = Allo(payable(address(proxy)));
 
         // Initialize
         vm.prank(protocolDeployer);
-        allo.initialize(
-            protocolDeployer,
-            address(registry),
-            payable(treasury),
-            percentFee,
-            baseFee,
-            forwarder
-        );
+        allo.initialize(protocolDeployer, address(registry), payable(treasury), percentFee, baseFee, forwarder);
 
         // Deploy base strategy
         strategy_directAllocation = new DirectAllocationStrategy(address(allo));
@@ -72,19 +56,12 @@ contract Setup is Actors {
         token = ERC20(address(new FuzzERC20()));
 
         // Create profile for each medusa sender
-        for (uint i; i < DEFAULT_MEDUSA_SENDER.length; i++) {
+        for (uint256 i; i < DEFAULT_MEDUSA_SENDER.length; i++) {
             bytes32 _id = registry.createProfile(
-                0,
-                "a",
-                Metadata({protocol: i + 1, pointer: ""}),
-                DEFAULT_MEDUSA_SENDER[i],
-                new address[](0)
+                0, "a", Metadata({protocol: i + 1, pointer: ""}), DEFAULT_MEDUSA_SENDER[i], new address[](0)
             );
 
-            _addActorAndAnchor(
-                DEFAULT_MEDUSA_SENDER[i],
-                registry.getProfileById(_id).anchor
-            );
+            _addActorAndAnchor(DEFAULT_MEDUSA_SENDER[i], registry.getProfileById(_id).anchor);
         }
     }
 }
